[REFACTOR] Rename constants file to config
