+++ src/tools/graphicsDatabase.ts
@@ -1,6 +1,6 @@
 import { Immutable } from 'immer';
 import { AssetGraphicsDefinition, AssetId, GetLogger, PointTemplate, type AssetSourceGraphicsInfo, type GraphicsDefinitionFile, type GraphicsSourceDefinitionFile } from 'pandora-common';
-import { GENERATE_AVIF } from '../constants.ts';
+import { GENERATE_AVIF } from '../config.ts';
 import { AUTOMESH_TEMPLATES } from '../templates/automeshTemplates.ts';
 import { AVIF_SUFFIX } from './resources.ts';
